rune and raw rpc examples
