+++ examples/rune/u_track.ts
@@ -1,3 +1,10 @@
+/**
+ * @title Rune U Track
+ * @stability nearing
+ * @description Rune U Track
+ * @todo
+ */
+
 import { assert, assertEquals, assertInstanceOf } from "asserts"
 import { Rune } from "capi"
